Move renderRenderPropComponent into Dropdown
