+++ src/components/Dropdown/Dropdown.MenuContainer.jsx
@@ -30,7 +30,7 @@ import {
 import { MenuContainerUI } from './Dropdown.css.js'
 import classNames from 'classnames'
 import getShallowDiffs from '@helpscout/react-utils/dist/getShallowDiffs'
-import { renderRenderPropComponent } from '../../utilities/component'
+import { renderRenderPropComponent } from './Dropdown.utils'
 import { createUniqueIDFactory } from '../../utilities/id'
 import { getComputedClientRect } from './Dropdown.MenuContainer.utils'
